GH-372: add auto extending timeouts on fetch, search, and copy (#375)
* comment out idle test for now

* update version

* Turn off forced folder check on initial scan. Properly exit the read loop when closing the client connection

It was making user server startup take too long. It should be fine this way.

Also if we closed the client write stream, but did not exist the read loop we would still get messages from the client.

* The bug reports are all ancient history.

Also tracking them externally (github issues for now) and via branches in the future

* update requirements

* Implement auto-extending timeouts for copy, search, and fetch.

Also move default timeout to 120s.
Question is if the search is being locked out by a long running fetches, it will fail anyways. Maybe we should extend the timeout in the ‘wait for mbox ready’ loop as long as another auto-extending command is running?